backend/requirements.txt
fastapi[standard]
requests
python-dotenv
geocoder
google-genai
sentence-transformers
